refactor and update other links
