+++ apps/src/codebridge/utils/index.ts
@@ -1,5 +1,6 @@
 export * from './analyticsReporterHelper';
 export * from './editableFileType';
+export * from './createBlobUrlForFile';
 export * from './isDuplicateFileName';
 export * from './isDuplicateFolderName';
 export * from './isValidFileName';
